fix: just shut off that extra warnings of gcc :D

+++ include/ciphers/classical/affine.h
@@ -12,7 +12,7 @@ typedef struct {
 
 
 
-void affine_encrypt(const uchar_t* input, uchar_t* output , int length, const void* key);
+void affine_encrypt(const uchar_t* input, uchar_t* output , size_t length, const void* key);
 void affine_decrypt(const uchar_t* input, uchar_t* output, const void* key);
 void affine_set_key(void* key_struct, const char* key_str);
 void affine_free_key(void* key_struct);
